requirements.txt
# ==========================================================
# Slack
# ==========================================================
slack-bolt==1.18.1
slack-sdk==3.26.2

# ==========================================================
# HTTP / APIs externas
# ==========================================================
requests==2.31.0

# ==========================================================
# Configuração de ambiente
# ==========================================================
python-dotenv==1.0.1

# ==========================================================
# Utilidades
# ==========================================================
typing-extensions==4.9.0